pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "semantic-moments"
version = "0.1.0"
description = "Training-Free Motion Similarity via Third Moment Features"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
keywords = ["video", "motion", "similarity", "embeddings", "computer-vision"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.40.0",
    "pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "tqdm>=4.65.0",
]

[project.urls]
"Homepage" = "https://github.com/YOUR_USERNAME/semantic-moments"
"Paper" = "https://arxiv.org/abs/2602.09146"

[tool.setuptools.packages.find]
where = ["src"]
